requirements.txt
# ZAPBOT DEPENDECIES

selenium
ipykernel
pyyaml